# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/companies/{companyId}/applications": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "description": "VCP-controlled company identifier",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Applications owned by the requested company.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is missing, malformed, expired, or cannot be verified.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Verified subject has no active access to the requested company.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The configured identity verifier cannot establish trust.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List applications visible in a company",
        "tags": [
          "applications"
        ]
      },
      "post": {
        "operationId": "register",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "description": "VCP-controlled company identifier",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterApplicationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New or previously registered idempotent application result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request body failed schema validation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ValidationIssueResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is missing, malformed, expired, or cannot be verified.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Verified subject has no active access to the requested company.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The configured identity verifier cannot establish trust.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register an application idempotently",
        "tags": [
          "applications"
        ]
      }
    },
    "/companies/{companyId}/applications/{applicationId}/assessments": {
      "post": {
        "operationId": "submit",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "description": "VCP-controlled company identifier",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAssessmentRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Queued assessment, or the original assessment for an idempotent retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request body failed schema validation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ValidationIssueResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is missing, malformed, expired, or cannot be verified.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Verified subject has no active access to the requested company.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The configured identity verifier cannot establish trust.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Queue an asynchronous assessment idempotently",
        "tags": [
          "assessments"
        ]
      },
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "description": "VCP-controlled company identifier",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Assessments for the requested company application, oldest first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssessmentResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is missing, malformed, expired, or cannot be verified.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Verified subject has no active access to the requested company.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The configured identity verifier cannot establish trust.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List assessments for an application",
        "tags": [
          "assessments"
        ]
      }
    },
    "/companies/{companyId}/applications/{applicationId}/assessments/{assessmentId}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "name": "assessmentId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "description": "VCP-controlled company identifier",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Current asynchronous lifecycle state and terminal result when available.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is missing, malformed, expired, or cannot be verified.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Verified subject has no active access to the requested company.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No assessment with this identifier exists in the requested company.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The configured identity verifier cannot establish trust.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get asynchronous assessment status and result",
        "tags": [
          "assessments"
        ]
      }
    }
  },
  "info": {
    "title": "Vibe Coding Platform Control Plane",
    "description": "Architecture-spike contract for company-scoped application registration and asynchronous assessment.",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Verified identity-provider access token. The local spike token is not for production use."
      }
    },
    "schemas": {
      "HttpErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401,
            "description": "HTTP status code."
          },
          "message": {
            "type": "string",
            "example": "Verified identity is required",
            "description": "Safe, human-readable failure summary."
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
      "ApplicationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "VCP application identifier."
          },
          "companyId": {
            "type": "string",
            "description": "Owning VCP company identifier.",
            "example": "company-a"
          },
          "name": {
            "type": "string",
            "description": "Company-visible application name.",
            "example": "Readiness portal"
          },
          "repositoryUrl": {
            "type": "string",
            "format": "uri",
            "description": "Approved source repository URL."
          },
          "idempotencyKey": {
            "type": "string",
            "description": "Idempotency key accepted during registration."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "UTC registration timestamp."
          }
        },
        "required": [
          "id",
          "companyId",
          "name",
          "repositoryUrl",
          "idempotencyKey",
          "createdAt"
        ]
      },
      "RegisterApplicationRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Company-visible application name.",
            "example": "Readiness portal",
            "minLength": 1,
            "maxLength": 120
          },
          "repositoryUrl": {
            "type": "string",
            "description": "HTTPS URL of the approved source repository.",
            "example": "https://github.com/example/readiness-portal",
            "format": "uri"
          },
          "idempotencyKey": {
            "type": "string",
            "description": "Caller-generated key that makes retries return the original application.",
            "example": "register-readiness-portal-v1",
            "minLength": 8,
            "maxLength": 100
          }
        },
        "required": [
          "name",
          "repositoryUrl",
          "idempotencyKey"
        ]
      },
      "ValidationIssueResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "too_small",
            "description": "Zod validation issue code."
          },
          "path": {
            "example": [
              "idempotencyKey"
            ],
            "description": "Request-body property path. Numeric array indexes are serialized as values.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "message": {
            "type": "string",
            "example": "Too small: expected string to have >=8 characters"
          }
        },
        "required": [
          "code",
          "path",
          "message"
        ]
      },
      "SubmitAssessmentRequest": {
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "description": "Caller-generated key unique to this application assessment request.",
            "example": "assess-main-commit-v1",
            "minLength": 8,
            "maxLength": 100
          }
        },
        "required": [
          "idempotencyKey"
        ]
      },
      "AssessmentResultResponse": {
        "type": "object",
        "properties": {
          "profile": {
            "type": "string",
            "example": "placeholder-web-application",
            "description": "Detected application profile. This spike currently returns a placeholder profile."
          },
          "findings": {
            "example": [],
            "description": "Machine-produced readiness findings; empty when no findings are produced.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "profile",
          "findings"
        ]
      },
      "AssessmentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Assessment identifier used for status polling."
          },
          "companyId": {
            "type": "string",
            "description": "Owning VCP company identifier.",
            "example": "company-a"
          },
          "applicationId": {
            "type": "string",
            "format": "uuid",
            "description": "Application being assessed."
          },
          "idempotencyKey": {
            "type": "string",
            "description": "Idempotency key accepted for this assessment."
          },
          "correlationId": {
            "type": "string",
            "description": "Correlation identifier retained from submission through worker audit events.",
            "example": "01J5VCPASSESSMENT"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "example": "queued",
            "description": "Asynchronous lifecycle state. Completed and failed are terminal."
          },
          "attempts": {
            "type": "number",
            "minimum": 0,
            "example": 0,
            "description": "Number of worker claims, including recovery claims after a stale lock."
          },
          "result": {
            "description": "Present only when status is completed.",
            "allOf": [
              {
                "$ref": "#/components/schemas/AssessmentResultResponse"
              }
            ]
          },
          "error": {
            "type": "string",
            "description": "Safe failure summary present only when status is failed.",
            "example": "Assessment worker could not process the request"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "UTC submission timestamp."
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "description": "UTC timestamp of the most recent worker claim."
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "description": "UTC timestamp when the assessment entered a terminal state."
          }
        },
        "required": [
          "id",
          "companyId",
          "applicationId",
          "idempotencyKey",
          "correlationId",
          "status",
          "attempts",
          "createdAt"
        ]
      }
    }
  }
}